requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM & Embeddings
openai>=1.0.0
llama-index>=0.9.0
llama-index-embeddings-openai-like>=0.1.0
llama-index-vector-stores-faiss>=0.1.0
llama-index-retrievers-bm25>=0.1.0
llama-index-readers-file>=0.1.0

# Vector Store
faiss-cpu>=1.7.4

# HTTP Clients
httpx>=0.25.0
requests>=2.31.0

# UI
gradio>=4.0.0

# Utilities
python-multipart>=0.0.6
tiktoken>=0.5.0

# Optional: PDF support
PyMuPDF>=1.23.0